requirements.txt
notebook
python-dateutil
pandas
matplotlib
plotly